post-install: Improve scripts

+++ post-install.sh
@@ -3,28 +3,18 @@
 set -e
 
 BASEPATH="${BASEPATH:-/tmp}"
-MODE="${MODE:-gui}"
-
-is_container() {
-  if cat /proc/self/cgroup | grep -q "docker/"; then
-    return 0
-  fi
-
-  if cat /proc/self/cgroup | grep -q "kubepods/"; then
-    return 0
-  fi
-
-  if [ -n "$CONTAINER" ]; then
-    return 0
-  fi
-
-  return 1
-}
+CONTAINER=${CONTAINER:-0}
+GUI_ENABLED="${GUI_ENABLED:-1}"
+HARDWARE=${HARDWARE:-1}
 
 cd "$BASEPATH"
 
 apt-get update
 apt-get upgrade -y
+
+################
+# Installation #
+################
 
 apt-get install -y \
   apt-transport-https \
@@ -52,7 +42,7 @@ apt-get install -y \
   wget \
   zsh
 
-if ! is_container; then
+if [ $HARDWARE -ne 0 ]; then
   apt-get install -y \
     btrfs-progs \
     cryptsetup \
@@ -63,18 +53,18 @@ if ! is_container; then
     usbutils \
     vbetool
 
-  if lspci | grep -q "Network controller"; then
-    apt-get install -y rfkill wireless-tools wpasupplicant
-
-    if [ "$MODE" == "gui" ]; then
-      apt-get install -y wicd-gtk
-    elif [ "$MODE" == "text" ]; then
-      apt-get install -y wicd-curses
+  if [ $GUI_ENABLED -eq 0 ]; then
+    if lspci | grep -q "Network controller"; then
+      apt-get install -y rfkill wicd-curses wireless-tools wpasupplicant
     fi
-  fi
+  else
+    if lspci | grep -q "Network controller"; then
+      apt-get install -y rfkill wicd-gtk wireless-tools wpasupplicant
+    fi
 
-  if lsmod | grep -q "bluetooth"; then
-    apt-get install -y blueman
+    if lsmod | grep -q "bluetooth"; then
+      apt-get install -y blueman
+    fi
   fi
 fi
 
@@ -91,7 +81,7 @@ chmod +x /bin/busybox
 
 # Docker
 
-if ! is_container; then
+if [ $CONTAINER -ne 0 ]; then
   if [ ! -f containerd.io_1.2.5-1_amd64.deb ]; then
     wget 'https://download.docker.com/linux/debian/dists/buster/pool/stable/amd64/containerd.io_1.2.5-1_amd64.deb'
   fi
@@ -113,7 +103,7 @@ fi
 
 # Docker Compose
 
-if ! is_container; then
+if [ $CONTAINER -ne 0 ]; then
   if [ ! -f docker-compose-1.24.0-Linux-x86_64 ]; then
     wget -O docker-compose-1.24.0-Linux-x86_64 'https://github.com/docker/compose/releases/download/1.24.0/docker-compose-Linux-x86_64'
   fi
@@ -147,7 +137,7 @@ rm -rf golangci-lint-1.16.0-linux-amd64
 
 # Hard Disk Sentinel
 
-if ! is_container; then
+if [ $HARDWARE -ne 0 ]; then
   if [ ! -f hdsentinel-017-x64.gz ]; then
     wget 'https://www.hdsentinel.com/hdslin/hdsentinel-017-x64.gz'
   fi
@@ -235,69 +225,72 @@ fi
 cp -f urchin-v0.1.0-rc3 /usr/bin/urchin
 chmod +x /usr/bin/urchin
 
-case "$MODE" in
-  "gui" )
+if [ $GUI_ENABLED -ne 0 ]; then
+  apt-get install -y \
+    chromium \
+    conky \
+    evince \
+    gimp \
+    inkscape \
+    transmission \
+    vlc \
+    xfce4 \
+    xfce4-goodies
+
+  if [ $HARDWARE -ne 0 ]; then
     apt-get install -y \
       alsa-utils \
-      chromium \
-      conky \
       cups \
-      evince \
-      gimp \
-      inkscape \
       simple-scan \
       system-config-printer \
-      transmission \
-      vlc \
-      xcalib \
-      xfce4 \
-      xfce4-goodies
+      xcalib
+  fi
 
-    # Paper Theme
+  # Paper Theme
 
-    if [ ! -f paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb ]; then
-      wget 'https://launchpadlibrarian.net/383884507/paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb'
-    fi
+  if [ ! -f paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb ]; then
+    wget 'https://launchpadlibrarian.net/383884507/paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb'
+  fi
 
-    dpkg -i paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb ||
-      apt-get install -fy
+  dpkg -i paper-icon-theme_1.5.721-201808151353~daily~ubuntu18.04.1_all.deb ||
+    apt-get install -fy
 
-    if [ ! -f paper-gtk-theme.tar.gz ]; then
-      wget -O paper-gtk-theme.tar.gz 'https://github.com/snwh/paper-gtk-theme/archive/master.tar.gz'
-    fi
+  if [ ! -f paper-gtk-theme.tar.gz ]; then
+    wget -O paper-gtk-theme.tar.gz 'https://github.com/snwh/paper-gtk-theme/archive/master.tar.gz'
+  fi
 
-    tar -xf paper-gtk-theme.tar.gz
-    (cd paper-gtk-theme-master && ./install-gtk-theme.sh)
-    rm -rf paper-gtk-theme-master
+  tar -xf paper-gtk-theme.tar.gz
+  (cd paper-gtk-theme-master && ./install-gtk-theme.sh)
+  rm -rf paper-gtk-theme-master
 
-    # ST
+  # ST
 
-    if [ ! -f st-0.8.2.tar.gz ]; then
-      wget 'https://dl.suckless.org/st/st-0.8.2.tar.gz'
-    fi
+  if [ ! -f st-0.8.2.tar.gz ]; then
+    wget 'https://dl.suckless.org/st/st-0.8.2.tar.gz'
+  fi
 
-    if [ ! -f https://st.suckless.org/patches/alpha/st-alpha-0.8.2.diff ]; then
-      wget 'https://st.suckless.org/patches/alpha/st-alpha-0.8.2.diff'
-    fi
+  if [ ! -f https://st.suckless.org/patches/alpha/st-alpha-0.8.2.diff ]; then
+    wget 'https://st.suckless.org/patches/alpha/st-alpha-0.8.2.diff'
+  fi
 
-    if [ ! -f st-clipboard-0.8.2.diff ]; then
-      wget 'https://st.suckless.org/patches/clipboard/st-clipboard-0.8.2.diff'
-    fi
+  if [ ! -f st-clipboard-0.8.2.diff ]; then
+    wget 'https://st.suckless.org/patches/clipboard/st-clipboard-0.8.2.diff'
+  fi
 
-    apt-get install -y gcc libx11-dev libxft-dev libxext-dev make
-    tar -xf st-0.8.2.tar.gz
+  apt-get install -y gcc libx11-dev libxft-dev libxext-dev make
+  tar -xf st-0.8.2.tar.gz
 
-    (
-      cd st-0.8.2
-      git apply $(find .. -name "st-*.diff")
-      sed -i "s/defaultbg = 257/defaultbg = 0/" config.def.h
-      sed -i "s/alpha = 0xcc/alpha = 0xdd/" config.def.h
-      make clean install
-    )
+  (
+    cd st-0.8.2
+    git apply $(find .. -name "st-*.diff")
+    sed -i "s/defaultbg = 257/defaultbg = 0/" config.def.h
+    sed -i "s/alpha = 0xcc/alpha = 0xdd/" config.def.h
+    make clean install
+  )
 
-    rm -rf st-0.8.2
+  rm -rf st-0.8.2
 
-    cat <<EOF > /usr/share/applications/simple-terminal.desktop
+  cat <<EOF > /usr/share/applications/simple-terminal.desktop
 [Desktop Entry]
 Name=st
 GenericName=Terminal
@@ -311,31 +304,31 @@ Categories=System;TerminalEmulator;
 Keywords=shell;prompt;command;commandline;cmd;
 EOF
 
-    # Telegram
+  # Telegram
 
-    if [ ! -f tsetup.1.6.7.tar.xz ]; then
-      wget 'https://updates.tdesktop.com/tlinux/tsetup.1.6.7.tar.xz'
-    fi
+  if [ ! -f tsetup.1.6.7.tar.xz ]; then
+    wget 'https://updates.tdesktop.com/tlinux/tsetup.1.6.7.tar.xz'
+  fi
 
-    tar -xf tsetup.1.6.7.tar.xz -C /opt/
-    ln -sf /opt/Telegram/Telegram /usr/bin/telegram
+  tar -xf tsetup.1.6.7.tar.xz -C /opt/
+  ln -sf /opt/Telegram/Telegram /usr/bin/telegram
 
-    # Vim
+  # Vim
 
-    if [ ! -f vim-8.1.tar.bz2 ]; then
-      wget 'ftp://ftp.vim.org/pub/vim/unix/vim-8.1.tar.bz2'
-    fi
+  if [ ! -f vim-8.1.tar.bz2 ]; then
+    wget 'ftp://ftp.vim.org/pub/vim/unix/vim-8.1.tar.bz2'
+  fi
 
-    apt-get purge -fy vim-tiny
+  apt-get purge -fy vim-tiny
 
-    apt-get install -y \
-      gcc libncurses-dev libx11-dev libxpm-dev libxt-dev libxtst-dev make
+  apt-get install -y \
+    gcc libncurses-dev libx11-dev libxpm-dev libxt-dev libxtst-dev make
 
-    tar -xf vim-8.1.tar.bz2
-    (cd vim81 && ./configure --with-features=huge && make && make install)
-    rm -rf vim81
+  tar -xf vim-8.1.tar.bz2
+  (cd vim81 && ./configure --with-features=huge && make && make install)
+  rm -rf vim81
 
-    cat <<EOF > /usr/share/applications/vim.desktop
+  cat <<EOF > /usr/share/applications/vim.desktop
 [Desktop Entry]
 Name=Vim
 GenericName=Text Editor
@@ -350,25 +343,23 @@ Categories=Utility;TextEditor;Development;
 Keywords=Text;editor;
 EOF
 
-    ;;
+elif [ $GUI_ENABLED -eq 0 ]; then
+  # Vim
 
-  "text" )
-    # Vim
+  if [ ! -f vim-8.1.tar.bz2 ]; then
+    wget 'ftp://ftp.vim.org/pub/vim/unix/vim-8.1.tar.bz2'
+  fi
 
-    if [ ! -f vim-8.1.tar.bz2 ]; then
-      wget 'ftp://ftp.vim.org/pub/vim/unix/vim-8.1.tar.bz2'
-    fi
+  apt-get purge -fy vim-tiny
+  apt-get install -y gcc libncurses-dev make
+  tar -xf vim-8.1.tar.bz2
+  (cd vim81 && ./configure && make && make install)
+  rm -rf vim81
+fi
 
-    apt-get purge -fy vim-tiny
-    apt-get install -y gcc libncurses-dev make
-    tar -xf vim-8.1.tar.bz2
-    (cd vim81 && ./configure && make && make install)
-    rm -rf vim81
-
-    ;;
-esac
-
-# Cleaning
+############
+# Cleaning #
+############
 
 apt-get autoremove -y > /dev/null
 # shellcheck disable=SC2046
